Split the climate features into their own node.
